requirements.txt
attrs==22.2.0
aws-cdk-lib==2.103.0
boto3==1.26.55
botocore==1.29.55
cattrs==22.2.0
cdk-nag==2.27.173
constructs==10.1.228
exceptiongroup==1.1.0
jmespath==1.0.1
jsii==1.91.0
publication==0.0.3
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14